+++ crates/version-spec/tests/resolved_spec_test.rs
@@ -1,5 +1,5 @@
 use semver::Version;
-use version_spec::{SemVer, VersionSpec};
+use version_spec::{CalVer, SemVer, VersionSpec};
 
 mod resolved_spec {
     use super::*;
@@ -38,6 +38,16 @@ mod resolved_spec {
         assert_eq!(
             VersionSpec::parse("1.2.3").unwrap(),
             VersionSpec::Semantic(SemVer(Version::new(1, 2, 3)))
+        );
+
+        // calver
+        assert_eq!(
+            VersionSpec::parse("2024-02").unwrap(),
+            VersionSpec::Calendar(CalVer(Version::new(2024, 2, 0)))
+        );
+        assert_eq!(
+            VersionSpec::parse("2024-2-26").unwrap(),
+            VersionSpec::Calendar(CalVer(Version::new(2024, 2, 26)))
         );
     }